ws_handler/requirements.txt
PyJWT>=2.8.0,<3.0
boto3>=1.34.0
